Move io_file from tests to model
