+++ flowserv/tests/files.py
@@ -8,14 +8,7 @@
 
 """Helper classes and methods for unit tests that perform I/O operations."""
 
-import json
-
-from io import BytesIO
-from typing import Dict, List, Optional, Union
-
-from flowserv.volume.base import IOHandle, IOBuffer
-
-import flowserv.util as util
+from flowserv.volume.base import IOHandle
 
 
 class FileStorage(object):
@@ -40,25 +33,3 @@ class FileStorage(object):
     def save(self, filename: str):
         """Write file to disk or to a given IO buffer."""
         self.file.store(filename)
-
-
-def MemStore(env: Dict):
-    """Create an instance of the buckect store with a memory bucket."""
-    from flowserv.model.files.bucket import BucketStore
-    from flowserv.model.files.mem import MemBucket
-    return BucketStore(bucket=MemBucket())
-
-
-# -- Helper Functions ---------------------------------------------------------
-
-
-def io_file(data: Union[List, Dict], format: Optional[str] = None) -> IOBuffer:
-    """Write simple text to given bytes buffer."""
-    buf = BytesIO()
-    buf.seek(0)
-    if format is None or format == util.FORMAT_JSON:
-        buf.write(str.encode(json.dumps(data)))
-    else:
-        for line in data:
-            buf.write(str.encode('{}\n'.format(line)))
-    return IOBuffer(buf)
