+++ tests/controller/serial/engine/test_postproc_workflow.py
@@ -13,9 +13,9 @@ import pytest
 import time
 
 from flowserv.config import Config
+from flowserv.model.files import io_file
 from flowserv.service.local import LocalAPIFactory
 from flowserv.service.run.argument import serialize_arg, serialize_fh
-from flowserv.tests.files import io_file
 from flowserv.tests.service import (
     create_group, create_user, create_workflow, start_run, upload_file
 )
