+++ tests/controller/serial/engine/test_sync_controller.py
@@ -11,11 +11,9 @@
 import os
 import pytest
 
+from flowserv.model.files import io_file
 from flowserv.service.run.argument import serialize_arg, serialize_fh
-from flowserv.tests.files import io_file
-from flowserv.tests.service import (
-    create_group, create_user, create_workflow, start_run, upload_file
-)
+from flowserv.tests.service import create_group, create_user, create_workflow, start_run, upload_file
 
 
 import flowserv.model.workflow.state as st
