+++ tests/view/test_view_files.py
@@ -8,12 +8,11 @@
 
 """Unit tests for the user resources view."""
 
-from flowserv.config import Config
-from flowserv.model.files.fs import FileSystemStore
+from flowserv.model.files import io_file
 from flowserv.model.group import WorkflowGroupManager
-from flowserv.tests.files import io_file
 from flowserv.view.files import UploadFileSerializer
 from flowserv.view.validate import validator
+from flowserv.volume.fs import FileSystemStorage
 
 import flowserv.tests.model as model
 import flowserv.view.files as labels
@@ -21,11 +20,10 @@ import flowserv.view.files as labels
 
 def test_file_listing_serialization(database, tmpdir):
     """Test serialization of file handles."""
-    config = Config().basedir(tmpdir)
     view = UploadFileSerializer()
     filename = 'data.json'
     with database.session() as session:
-        manager = WorkflowGroupManager(session=session, fs=FileSystemStore(config))
+        manager = WorkflowGroupManager(session=session, fs=FileSystemStorage(basedir=tmpdir))
         user_id = model.create_user(session, active=True)
         workflow_id = model.create_workflow(session)
         group_id = model.create_group(session, workflow_id, users=[user_id])
